Add Obsidian -> Lava Bucket (#285)
* Add Obsidian -> Lava Bucket

This is a useful feature for those absolute degenerates that dont know how to make a cobble gen, who doesnt know how to make a cobblegen in 2021? anyway lets not bully those bakas anymore.

It can be used to covert obsidian back into lava whenever they do mess up some how. I wrote this for 1.17 so it likely isnt that backwards compatible but whatever, Suck it old verisons

* Lived 17 years, Still can't english

* Allow multiple buckets

* Config Option & Crouching conversioni