This PR follows the PR: IridiumSkyblock#368
Since we can put water below layer 0, we can also be trying to make an underwater island, the concern is that if we put a dome, it is filled with water and it is not playable! This PR allows the blow to ignore or not the installation of air block.